Add work around & soft fail for poo#116191
SLES+HA migrations from 15-SP3 to 15-SP4 & 15-SP5 on ppc64le fail on the
ha/check_after_reboot test module because the UUID configured for the
cluster_md MD RAID device in /etc/mdadm.conf in the qcow2 images
from the 15-SP3 installations is different than the actual block devices
UUID after the migration. This commit works around the issue and soft
fail the test when the issue is detected.