Refactor powerman module
This started as an investigation for the weird last validation from `powerman
-c`. it turned out that this was not an issue as the device used on the test
it doesnt supported. So the question was if it should be removed or
improved in one or the other way.

So as part of that improvement, the module checks for the capabilities of the
divice given by the specification files. In additional, the validation of the
commands are replaced by `validate_script_output` and some info added to the
OpenQA UI.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>