Add workaround for bsc#1184679 on wizard_hana_install
Module sles4sap/wizard_hand_install is currently failing on some
specific bare metal Dell servers on BACKEND=ipmi with a crash on the
Gnome session that takes the SUT back to the displaymanager login
screen. Since the related bug hasn't been active in a while, this adds a
workaround and a softfail.