Move xmlresult specific subs outside of the slurm test module
Doing so, and move them to the lib we can reuse the subroutines in other
modules as well. This is the first step of restructuring the slurm module.
For now the subroutines do not consider complicated scheduling. That means
that it hasnt tested in a scenario where two modules want to extract test
results into an xml file. But with current changes I drive the functionality
in this direction. For instance the *pars_results* can define a different
xmlfile to store test results per module.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>