Fix xml notation for HPC profiles
`t` is a new notation which is supported by sle15sp3+ so using it for older
versions makes the installation complain.
In addition I have fixed package conflict between those version regarding python
and package requirements between sle versions. `shim` is available to all the x86_64
however that is not true for aarch64 for sle15sp2 and below.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>