Stabilize mpi tests using LF on type_string
This changes is to stabilize the expected behavior with explicit newline in
the passed string.
Make sure that the `type_string` does get a new line at the end of the
typing and fix the timeout issue increasing the `ready_time` of
`wait_boot_textmode` which in some systems takes longer time to show the
prompt.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>