Move some yast maintenance update schedules
We created a directory called sle_15_sp3 directory some time ago, with
the idea to be used later with flows. We there a schedule file for a test suite
running on sle15 sp3. But currently that does not make anymore sense, as we
end-up with schedules for sle15 sp4 or even sle 12 in a directory callled sle_15_sp3.
Besides, other similar schedule files were placed under maintenance/.
This change moves the lost files where it makes more sense.