Install *-hpc-gnu packages on HPC compute nodes
Some time ago I implemented a NFS sharing between management and compute
nodes. This had the advantage that we had to install the binaries just once
and make them available immediately to the other nodes. Only requirement was
that the compute nodes were needed to install some dependencies. It turned out
that this might not an optimal solution for those binaries as the dependencies
are likely to cause a hell among the various products. As it happens with the
latest builds on sle15sp5 after some changes on openmpi.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>