Add soft fail workaround for bsc#1204528 bsc#1204978
Since gh#yast/yast-iscsi-client#121, the packages open-iscsi and iscsiuio
have been dropped as dependencies for yast2-iscsi-client, requiring
these packages to be explicitly installed to properly use the yast2
iscsi client module. This commit attempts the yast2-iscsi-client
installation with --recommends, and then verifies whether the dropped
dependency packages were installed; if not installed, a soft fail is
recorded and the dependencies are installed explicitly. A check for the
release notes of SLES 15-SP5 was also added when running on SLES 15-SP5:
if the documentation bug is present in the release notes, then the soft
fail will not be recorded.