publiccloud : Optimize registercloud log upload
Before this patch upload of registercloud log was
spreaded in several places. Which makes us uploading
file with same content several times. In some cases we
even uploading empty files.
This patch brings order into this process. From now on
we will upload regsitercloud only when there is test
failure AND file is not empty.