wicked: Add authoritative to dhcp-server config
This is the second approach for bsc#1204604 the first
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/15868
didn't fixed it.

Adding `authoritative;` let the DHCP server also answer on REQUEST, if the
request has a IP which doesn't belong to the own pool. This is a common
configuration, if you have only one DHCP-Server in a broadcast domain.