Set patterns to default for HPC installations of system role
HPC test suites which do a regular installation do not need to use selective
patterns as the purpose is to verify a happy path with the recommended
configuration.
This change is solving also another problem on the way the Openqa operation
works to select patterns in loop. The result is that the test does not install
the given patterns because the later deselection removes required pattern as a
dependency and causes unexpected fails (missing curl, in particular case).

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>