Redefine the script_sudo marker
Use `hashed_string` with the same signature which is used on
assert_script_sudo. The reason is that the same marker should be used to
assert serial output. Otherwise the returned wait_serial will not match in the
caller of the function, which uses this value with inner defined one.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>