Simplify default instance type
The default instance type evaluation doesn't work for ARM instances and
is needlessly complicated. This commit changes the behaviour to always
returns a `large` instance of the corresponding family (t2 for x86_64
and a1 for arm64), which simplifies the workflow considerably while
increasing the running costs only by a negligible amount.