Try SemVer before string cmp in check_version
By default, check_version was switching to string comparisson
if the Version module failed to parse the version strings. This
commit adds a second try to use the SemVer module before
resorting to string comparisson, as it is capable of parsing
more version strings. It will not be used on OS vcmp, as
OS vcmp relies on string comparisson.