Improve results parsing in ha/pacemaker_cts_cluster_exerciser
Current version of the parser can wrongfully pick results like `test
name\t failed` which are processed later in the output due to a small
issue in the awk command getting results from the log. This commit fixes
the awk command while also re-writing some of the commands so the fit in
under 100 characters per line.