Refactor mpi to apply c++ testing
That was left out in first place and needed to added. Firstly, the module
which was suppose to run c++ it was actually run simply c code. The
particularity with the sample code is that you need 2 nodes to run it. So that
wouldnt work for a one-standalone node.

I also exercise variations of the mpirun.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>