podman_pods: set replica count to 1 in hello-kubic.yaml
Since version 4.4.0, podman will limit the number of replicas to 1 when using
podman play kube. Thus it makes no sense anymore to have it set to 3 as that
avoids additional branches depending on the podman version. (Also, supporting
replicas was never really intended upstream and thus we shouldn't have tested it
anyway.)