Implement test for HPC warewulf4
A new test which will run one controller node and two compute nodes without
supportserver. An important aspect of this test is the controller VM which
requires two network interfaces. The compute nodes should run on this network
once they boot. Another is that the compute nodes are obliged to wait the
warewulf server to be ready before they boot. Because this is tricky on
OpenQA, the solution is keep the nodes reboot for some time. This
implementation has a defined number of reboot and likely it would be nice to
improve that loop to exit once it sees the booting process.

The testing is considered successful once it validates that controller can
connect remotely and the installed OS is confirmed.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>