Remove workaround of 'cannot load initrd' from pvm installation
we have issue 'cannot load initrd' since quite long time. The real root
cause is on qanet which has problem with exported directory mounts
sometimes.

So remove previous workaround for trying to reset at GRUB so that we can
save a lot of time for assert_screen and speed up the whole installation progress