Add HANA master and secondary instance tests
This commit creates the Azure PoC for SAP HANASR

This branch is a result of cherry-picking changes made by Lumir and Josef in
HANAsr_staging. The original branch was harder to merge because changes from
master were merged into it a long time ago, instead of having the branch rebase
onto master.

- Related tickets: TEAM-6627, TEAM-7430
- Verification run: http://openqaworker15.qa.suse.cz/tests/143171

Co-authored-by: Lumir Palovsky <lpalovsky@suse.com>
Co-authored-by: Josef Pupava <jpupava@suse.com>
Co-authored-by: Anton Pappas <apappas@suse.com>