Refactor spack to align with source code which is selected by get_mpi_src
boost is a special case which requires LD_LIBRARY_PATH which has been removed
and spack is not setting during its installation. A workaround causes problems
with the system packages and its fix makes the test complicated. Instead,
some other basic source code can avoid this, as it should be able to be
compiled and run without additional hacks. Although this does not decrease the
execution time. Meaning that spack still takes some time to install its
dependencies, including the openssl which was causing the issues in the
previous implementation.

This PR refactor the test to adjust the flow in accordance the source code which
will be selected by `get_mpi_src()` after removing the HPC_LIB. This variable
points to boost until now. Without the default source code will be selected.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>

Avoid conflicts between spack and system packages

spack tries to install libopenssl-devel if it is not installed and it does so
by fetching the external source. This breaks the compatibility with the
openssl installed in the system. loading the spack modules between the
underlined mpi commands avoids to mess with other system invocations.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>