Add users in slurm compute nodes
Users should appear among all the nodes of the cluster and their uid/gid
should match. It seems that the uid/gid are assigned correctly.

This used to work without the creation of the users on compute nodes for
SLE15SP4 and below but apparently it shouldnt. This commit just makes things
right.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>