extra_tests_kernel: Increase runtime and handle timeouts for ARM
On Aarch64 naptime.bit is terminated before attaching any
probes. Doubling the timeout prevents this on some tests, but not all.

This doubles the runtime because scripts run until they receive
SIGINT. Most likely doubling the timeout/runtime again will still not
prevent timeouts and is a waste.

So we handle timeouts as informational.