Let warewulf controller boot when controller is ready
Add mutex to remove controllers rebooting until the controller is
ready. Fix also the tear down operations which destroy the barriers.

Previously, as we implementated the loop, it could have finished before the
controller is ready to boot the nodes or the nodes would have to wait even if
the server was done.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>