Resurrect Azure Acceralated Network testing
The test module was not used for a while and needed to be updated
accordingly. Test suite creates 2 instances of type `Standard_D2_v4`
and measures network throuput using `iperf3`.
In order to install the package, we need to use opensuse repos or devel
repos in OBS as for SLE the `iperf` is provided by package hub and in
older versions only.

- ticket: [enable network performance test](https://progress.opensuse.org/issues/126224)