Use logs_from_installation_system on HPC installation
This PR allows logs_from_installation_system to collect logs from HPC autoyast
installation. I think this module should be applicable to be used without
modification but the installation console which used here was an
obstacle. When we use installation shell first we have to login and then we
have to exit, instead of return to it.
The changes do not provide a general solution but it just makes it work for
the HPC AY installation particular.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>