kernel/bpftrace: Try to increase timeout reliability
Instead of using the timeout utility to send an interrupt signal we
can append a timeout probe to the script we are executing.

If the interrupt signal gets ignored, which appears to happen under
stress, it is hard to say if it is a bug or not. If the interval probe
fails to fire or exit this seems like a more concrete bug.