Improve the usage of nfs when guest migration
Progress ticket: https://progress.opensuse.org/issues/133139
Utilizing NFS share storage in a virtualization environment:
either by creating a netfs storage pool on both source and destination hosts or by directly mounting the NFS directory on both hosts,
with the current testing setup involving only mounting the NFS directory on the destination host without mounting the NFS directory on the source host.
The purpose of doing this is to ensure that the same file system is accessed during guest migration.