Integrate Slurm versioned installation
Make possible to install and test latest version of the Slurm package. As for
now, Slurm is explicitly invoked. But now test should be able to get the
latest available versioned Slurm in the system, install it and run the tests.

Those changes are not apply to all the test cases at the moment. Verifications
run only against a basic scenario (see hpc_EPSILON) and the test requires a
new job variable, SLURM_VERSIONED_TEST.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>