azure: remove SKU from conditional check which type of image to use
Due to recent changes in Azure image upload logic ( move to image galaries )
we start to have jobs where ONLY OFFER is defined and default value used for SKU
This change was not addresed in get_image_id function where we were expecting BOTH
variables to be defined. This patch solving this problem.