Rework installation of k3s and kubectl
k3s server might not have been fully operational and were not waiting long
enough to let it properly start before usage.
Installation of k3s got also complicated and for certain distros we need
to pull it from upstream whereas for TW, MicroOS or the k3s deployment
script is already packaged. For k3s usage, we do not need to
specifically install `kubenernetes-client` as this handled in the k3s
deployment script.

Allow installation of kubernetes-client either by version that should be
part of the test suite setup or pull the newest available in the distro.

- tickets:
  - https://progress.opensuse.org/issues/132284
  - https://progress.opensuse.org/issues/134726