Do not install HANA from NFS. Copy media instead
This partially reverts 43558330cc2a61b28db0e65d8c40f60740d3d455, which
replaced the lib/sles4sap::copy_media method for a mount_media method,
enabling both HANA and NetWeaver installation tests to run the
installers directly from the NFS mount. In this commit, the copy_media
method has been added again to the library, and the HANA installation
test has been modified to use copy_media instead of mount_media.

General purpose of the commit is to be able to more easily identify
the root cause of some failures. With the current mount_media method, a
failing installation could have its root in the performance of the SUT,
network connection to the NFS server, NFS server performance and more.
If we copy the media first and then install from the local copy, we are
segmenting the steps, and we can more easily separate issues related to
the network or the NFS servers versus those related to the worker and
the SUT.

This also adds a unique NFSv4 client ID while mounting share.