xfstests: Add switchable non heartbeat mechanism
In some scenario test may need very good performance to test, and
heartbeat slow it down or just messed it up. The heartbeat mechanism
used to solve some hang and crash issue.
The non-heartbeat one use timeout to handle inspired by Josef's code
in his branch. And add some new features in it. To set a parameter
XFSTESTS_NO_HEARTBEAT=1 to close the heartbeat mechanism.
XFSTESTS_TIMEOUT to set the max time for a subtest. Test will send a
SIGTERM first when test run timeout and if command not stop it will
send a SIGKILL after 3 seconds.