Ansible deployment state
New status flag in the base class to record if Ansible has been executed or
not. Use the flag to only to call Ansible deregister when needed.
mr_test schedule to use qesap_cleanup in place of peering_destroy
Both peering_destroy and qesap_cleanup are doing exactly the same from
the point of view of the network peering to the IBSm.
qesap_cleanup has cluster destruction after that.