xfstests: Fix boot issue after a crash or hang
When test script know subtest in xfstests cause system crash/hang. It was
trigger reboot by power_action, and it will first find a root-console.
But crash/hang could be in two scenarios:
1. system hang in root console during run subtests;
2. system already crash and reboot by itself and waiting in bootloader screen.
The second scenario is not considered. Here to reboot "again" to keep logic and
real screen in the same page. And use 'alt-sysrq-b' to adapt both scenarios.