Replace sle-micro container with dedicated Warewulf container
As we want to be able to use the other images which the development team works on
we introduce a new test variable. This is intended to replace the current image
with home/eeich/sle-hpc-15-sp5-container/containers/suse/sle-hpc-node.
This one is built as base container for compute node deployment.
And this change is helping us to tests various images as such.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>