HPC: Explicitly install mariadb for slurm18 and sle15-sp1
All slurm versions install mariadb apart from slurm18 and slurm
base delivered for sle15sp1. That means for each case where there
is versioned slurm 18 mariadb must be installed explicitly. Moreover
slurm base for sle15 sp1 is exactly slurm 18 but since that isn't
a versioned slurm but a base one there must be another condition
which explicitly installs mariadb in that specific case. This condition
is kept separate with a comment to remove it once that code stream is
out of ltss