HPC: get_mpi refactor
As of now the get_mpi() is having somehow twisted logic where
everything is initially set to openmpi3. The patch is assuming the
base openmpi test setting is set to openmpi and then adjust version
depending of the sle hpc version. For sle12 it is openmpi1, for
sle15-sp1 it is openmpi2 and for sle15-sp2+ it is openmpi3