Refactor subroutine which creates xml result file
Move the parsing and the creation of the xml out of the SUT. As such we do not
repeat the script_run script which might make the runtime a bit faster
overall.
In addition the patch adds date and build elements in the xml as additional info
of the test results which it should help in the review of the tests.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>