LTP Whitelist: Add keep_fail attribute to whitelist JSON
A non-zero keep_fail attribute in a whitelist entry will prevent
matching failures from being whitelisted. Only a fail info box will
be created with the message provided by the whitelist entry to give
test reviewers reliable info about the failure, in particular
a bugref and symptoms so that different failures don't get mixed
together. This will be used for failures which need to be closely
tracked and cannot be whitelisted.