Use specified network interface in autoyast profile
Some SUT machines do not use default network interface,
for example, eth0 or em0. In order to have virtualization
bridge interface correctly configured, it is better to
use specified interface in bridge_ports instead of just
using a list of ports that can be available, which may
lead to ip lease inconsistency.