Force fencing module to run in root-console
Usually, `ha/fencing` module is followed by `boot/boot_to_desktop`,
which checks first for a grub screen. In this commit we force fencing
module to run in the root-console, so the node reboot goes directly from
the console to grub, even on systems installed with gnome. If not doing
this, systems installed with gnome will go the long way around of
attempting to match the grub needle on the displaymanager and fail after
a timeout.