Download SAP HANA install media from OSD instead of NFS
openQA can provide installation artifacts to the SUT from the test server.
Switching to that method from NFS allows us to run more tests as we are not
limited by the bandwidth of the NFS server but the connection of the openQA
server to the SUT. However the old method will be kept for now.

Additionally the server option is provided to the HANA installer as it is the
only one we are explicitly installing.

Related ticket: https://jira.suse.com/browse/TEAM-8573