HPC: Remove IMB runs form the basic mpi tests
The original idea was to have basic MPI tested with as basic as possible
MPI application. Over the time however the MPI tests got extended so
not only basic MPI app is run but also some other things like papi etc.
This isn't the best approach as the test code grows and more and more
ifs are added to account for various test paths.
This patch is the first to bring back the original idea, so the
simplicity of the code could be preserved. Other specifc tests should
be then moved to separate modules.

This patch removes IMB test. IMB tests rely on a regular SUSE package
and the test code re-use the test variable $mpi to install both: specific
MPI but also specific IMB version. While this seems like a good idea it
causes a trouble in case the rpms are being named differently. That
happened in case of openmpi which got renamed to openmpi1 but original
IMB rpm preserved openmpi in the name