Cleanup cache after unloading /usr/sbin/nscd profile
This is needed because the cache file has a newer timestamp than the
profile in /etc/apparmor.d/. Otherwise the next reload (in
lib/apparmortest.pm pre_run_hook()) would load the cached /usr/sbin/nscd
profile as created by aa-autodep.

The shipped profile that gets loaded (and is expected to be used) in the
aa-logprof test is named 'nscd' (without the path), therefore it won't
replace the /usr/sbin/nscd profile when it gets loaded.

The final result was that nscd was running under the wrong profile
(/usr/sbin/nscd from the cache), and therefore didn't create any log
entries for the 'nscd' profile.

With this fixed, we can also remove the workaround that was added to
cover this test bug in Tumbleweed for the last two years.

Note: I intentionally left some aa-status calls in the tests. They make
it much easier to verify that the expected profiles are loaded, and that
all processes run under the expected profile.

Fixes: https://bugzilla.opensuse.org/show_bug.cgi?id=1215856