HPC: Remove softfail for uid parameter on Slurm
Slurm 23_02 has removed the `--uid`. The check can be removed safely. In this
commit the `srun` commands which use `-N` parameter takes also the `-x` one
which set the nodes for exclusion from the node allocation of the job to
run. Those nodes are not configured properly and prints a warning in the stderr.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>