gnome/nautilus: close custom permission dialog
In recent GNOME versions, we need to press ESC once more to get
back to the file view. Pressing too many times ESC does not matter,
as we end up on the nautilus file list; this makes it easier for us
not having to worry about the GNOME version running.