Use firstboot module if ignition disk is not provided
Historically we have been using ignition disk to boot SLE Micro.
Since the jeos-firstboot wizard addition, we have now 2 options to
boot it. However, we were still using ignition as the default, and
forcing the variable FIRST_BOOT_CONFIG=wizard if we wanted to boot
with the wizard.

The problem was that if you don't provide either HDD_2=ignition or
FIRST_BOOT_CONFIG variables, the system will boot the wizard but
openQA will not load the corresponding test module.

This way, we don't need that variable any more to use the wizard
booting method, we just need to remove HDD_2 and NUMDISKS=2