PoC for codespaces configuration
Gitub Codespaces are dev environment within the browser.
They provides VM and in browser preconfigured VSCode.
The idea here is to define a preconfigured codespaces that
is targeting os-autoinst-distri-opensuse developers.
The VM image is configured by a Dockerfile and a script.
The environment is based on openSUSE TW. The image has packages suggested in
CONTRIBUTING.md plus other packages that are needed be able to run `make tests`.
The image is provided with non-root user `node`.
The proposed configuration also provide some pre-installed VSCode
plugins.