Combine download of tarball and unpacking
Due to load in osd workers, the `sles4sap/hana_install` test modules is
sporadically taking longer than 15 minutes unpacking the HANA installation
files tarball. This commit combines the download of the tarball asset with
its unpacking in a single pipeline with the 2700 second timeout which seems
to be enough to download the file; since tarball is not compressed, this
solution should accomplish getting the tarball unpacked into the SUT and
avoid the later timeout when trying to unpack a downloaded tarball.