Avoid die in check_takeover for uninitialized variables
Avoid an exception if get_hana_topology return an empty result. It could
happens if SAPHanaSR-showAttr is not succesful and output is not
parsable.
This new behavior is obtained avoiding the usage of local variables in
check_takeover.
Sleep 30 moved in a different position only to be used once for each
get_hana_topology and not to be called when there's no more retry to
run.
Add some unit testing for these two functions.